requirements.txt
anthropic>=0.39.0
requests>=2.32.0
pandas>=2.2.0
matplotlib>=3.9.0
numpy>=2.1.0
python-dotenv>=1.0.0
